docs: 주석 추가 및 파일 분리
